+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ------
 * [#965](https://github.com/Shopify/shopify-app-cli/pull/965): Remove --no-optional when using npm to create new project
+* [#958](https://github.com/Shopify/shopify-app-cli/pull/958): Split `connect` command into project-specific functionality
 
 Version 1.4.1
 ------